Test: Do not manually sign in before profile edit.
The workflow in `sign_in_as` is a dependency of the profile editing, and
it should not be. The signing in is also already under test elsewhere.

So, use the devise `sign_in` helper to directly sign the user in, taking care of that
machinery and breaking the dependency, letting this test focus solely on
profile editing.

cc: @nathanstruhs