Project: Rename missing image file.
The Rails asset pipeline precompiles assets and converts three-letter
filename extensions like jpeg to jpg. So, when an image tag specifies the
filename 'missing_image.jpeg', it doesn't match the filename of the precompiled
asset, and Rails fails to generate an appropriate URL.

Change the Paperclip default_url to use the '.jpg' extension. And, go ahead
and rename the actual asset file so it matches, just to avoid any confusion
about the magic matching.