State: Encapsulate logic for states with projects
The controller loads all the states, then delegates to the view
to decide on rendering only the states that have projects. This results
in rendering 51 partials that end up being empty.

Load only the appropriate states in the controller.
Define named scopes to encapsulate the controller loading logic.
Remove the logic from the partial.

References #307