StatesController: Remove unused routes, extract method into class
The StatesController only has two actions: index and show.
Update the routes to only allow index and show.

The StateController has a public method, accumulate_data. Extract
the method into a Hash method, add a test, fix the implementation,
and use it in the controller. Delete the public method.

Clean up the #show action.

References #312