AffiliationsController: Refine routes and lifecycle methods, add controller test
Reorder the before actions to check for authentication first, to enable early
redirecting of non-logged-in users.

Delete the show and new methods, views and routes - they are not used.

Clean up and fix the private controller methods.
Define Project authored_by? method and use it in place of explicit comparison
in the controller.

Add rspec devise and user helpers for signing in.
Use sequences for user factory attribute generation.

References #313