Organizations: Place chart loading js within event handler
When navigating from the index to the show view, the graph does not appear.
This is because the javascript is dumped at the end of the view, and
the dom isn't ready yet for binding the chart to the myChart element.

Place the chart binding and display in an event handler, bound to the
turbolinks:load event.

References #326, #275