Project: Don't display image if not persisted
The form is displaying a project cover photo if one exists. This is fine when editing
an existing record. But it causes an exception when:

Creating a new project
Specifying a file to attach for the cover photo
Submitting the form for the new project
Failing any validation

Upon re-rendering the project object has its cover_photo attachment, but the
attachment has not yet been persisted. So trying to render the attached image
via image_tag causes an exception.

Check to be sure that the project is persisted before trying to display its attached
cover photo.

References #295